+++ spec/rubocop/cop/performance/redundant_merge_spec.rb
@@ -37,6 +37,61 @@ describe RuboCop::Cop::Performance::RedundantMerge, :config do
     end
   end
 
+  context 'when internal to each_with_object' do
+    it 'autocorrects when the receiver is the object being built' do
+      source = ['foo.each_with_object({}) do |f, hash|',
+                '  hash.merge!(a: 1, b: 2)',
+                'end']
+      new_source = autocorrect_source(cop, source)
+
+      expect(new_source).to eq(['foo.each_with_object({}) do |f, hash|',
+                                '  hash[:a] = 1',
+                                '  hash[:b] = 2',
+                                'end'].join("\n"))
+    end
+
+    it 'autocorrects when the receiver is the object being built when ' \
+       'merge! is the last statement' do
+      source = ['foo.each_with_object({}) do |f, hash|',
+                '  some_method',
+                '  hash.merge!(a: 1, b: 2)',
+                'end']
+      new_source = autocorrect_source(cop, source)
+
+      expect(new_source).to eq(['foo.each_with_object({}) do |f, hash|',
+                                '  some_method',
+                                '  hash[:a] = 1',
+                                '  hash[:b] = 2',
+                                'end'].join("\n"))
+    end
+
+    it 'autocorrects when the receiver is the object being built when ' \
+       'merge! is not the last statement' do
+      source = ['foo.each_with_object({}) do |f, hash|',
+                '  hash.merge!(a: 1, b: 2)',
+                '  why_are_you_doing_this?',
+                'end']
+      new_source = autocorrect_source(cop, source)
+
+      expect(new_source).to eq(['foo.each_with_object({}) do |f, hash|',
+                                '  hash[:a] = 1',
+                                '  hash[:b] = 2',
+                                '  why_are_you_doing_this?',
+                                'end'].join("\n"))
+    end
+
+    it 'does not register an offense when merge! is being assigned inside ' \
+       'each_with_object' do
+      source = ['foo.each_with_object({}) do |f, hash|',
+                '  changes = hash.merge!(a: 1, b: 2)',
+                '  why_are_you_doing_this?',
+                'end']
+      inspect_source(cop, source)
+
+      expect(cop.offenses).to be_empty
+    end
+  end
+
   %w(if unless while until).each do |kw|
     context "when there is a modifier #{kw}, and more than 1 pair" do
       it "autocorrects it to an #{kw} block" do